refactor: fix reported style issues
