+++ esphome/components/hrxl_maxsonar_wr/hrxl_maxsonar_wr.cpp
@@ -9,7 +9,7 @@
 #include "esphome/core/log.h"
 
 namespace esphome {
-namespace hrxlmaxsonarwr {
+namespace hrxl_maxsonar_wr {
 
 static const char *const TAG = "hrxl.maxsonar.wr.sensor";
 static const uint8_t ASCII_CR = 0x0D;
@@ -23,7 +23,7 @@ void HrxlMaxsonarWrComponent::loop() {
   uint8_t data;
   while (this->available() > 0) {
     if (this->read_byte(&data)) {
-      buffer += (char) data;
+      buffer_ += (char) data;
       this->check_buffer_();
     }
   }
@@ -32,23 +32,17 @@ void HrxlMaxsonarWrComponent::loop() {
 void HrxlMaxsonarWrComponent::check_buffer_() {
   // Stop reading at ASCII_CR. Also prevent the buffer from growing
   // indefinitely if no ASCII_CR is received after MAX_DATA_LENGTH_BYTES.
-  if (this->buffer.back() == ASCII_CR ||
-      this->buffer.length() >= MAX_DATA_LENGTH_BYTES) {
-
-    ESP_LOGV(TAG, "Read from serial: %s", this->buffer.c_str());
-
-    if (this->buffer.length() == MAX_DATA_LENGTH_BYTES &&
-        this->buffer[0] == 'R' &&
-        this->buffer.back() == ASCII_CR) {
-
-      int millimeters = atoi(this->buffer.substr(1, MAX_DATA_LENGTH_BYTES - 2).c_str());
+  if (this->buffer_.back() == ASCII_CR || this->buffer_.length() >= MAX_DATA_LENGTH_BYTES) {
+    ESP_LOGV(TAG, "Read from serial: %s", this->buffer_.c_str());
+    if (this->buffer_.length() == MAX_DATA_LENGTH_BYTES && this->buffer_[0] == 'R' && this->buffer_.back() == ASCII_CR) {
+      int millimeters = atoi(this->buffer_.substr(1, MAX_DATA_LENGTH_BYTES - 2).c_str());
       float meters = float(millimeters) / 1000.0;
       ESP_LOGV(TAG, "Distance from sensor: %u mm, %f m", millis, meters);
       this->publish_state(meters);
     } else {
-      ESP_LOGW(TAG, "Invalid data read from sensor: %s", this->buffer.c_str());
+      ESP_LOGW(TAG, "Invalid data read from sensor: %s", this->buffer_.c_str());
     }
-    this->buffer.clear();
+    this->buffer_.clear();
   }
 }
 
@@ -59,5 +53,5 @@ void HrxlMaxsonarWrComponent::dump_config() {
   this->check_uart_settings(9600, 1, esphome::uart::UART_CONFIG_PARITY_NONE, 8);
 }
 
-}  // namespace hrxlmaxsonarwr
-}  // namspace esphome
+}  // namespace hrxl_maxsonar_wr
+}  // namespace esphome
